requirements.txt
arxiv
pdfplumber
sentence-transformers
numpy
faiss-cpu
networkx
requests
tqdm